fix: warnings on build due to renamed assemblies.
Had a conversation with QA around this and accidentally found a solution to the issue here: Unity doesn't warn about non-matching assembly names in builds when those assemblies are in an "Editor" subfolder. Turns out the "magic folder naming of old ages" still is in effect in some places. So until in some far future Unity fixes it, this is a workaround :)

Fixes #16